+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 4.0.0 - 2023-11-01
+
+### Changed
+
+- Requires `innmind/http-transport:~7.0`
+- Requires `innmind/http:~7.0`
+- Requires `innmind/filesystem:~7.1`
+- Requires `innmind/operating-system:~4.0`
+
+### Fixed
+
+- Removed files in directories were not removed from S3
+
 ## 3.2.1 - 2023-09-23
 
 ### Fixed